@@ -1,54 +1,72 @@
 {
-  "name": "angular-boilerplate",
-  "version": "0.0.0",
-  "private": true,
-  "scripts": {
-    "postinstall": "npx husky install",
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "watch": "ng build --watch --configuration development",
-    "test": "ng test",
-    "test:headless": "ng test --watch=false --browsers ChromeHeadless",
-    "lint": "ng lint"
-  },
-  "dependencies": {
-    "@angular/animations": "~15.0.4",
-    "@angular/common": "~15.0.4",
-    "@angular/compiler": "~15.0.4",
-    "@angular/core": "~15.0.4",
-    "@angular/forms": "~15.0.4",
-    "@angular/platform-browser": "~15.0.4",
-    "@angular/platform-browser-dynamic": "~15.0.4",
-    "@angular/router": "~15.0.4",
-    "rxjs": "~7.5.0",
-    "tslib": "^2.3.0",
-    "zone.js": "~0.12.0"
-  },
-  "devDependencies": {
-    "@angular-devkit/build-angular": "~15.0.5",
-    "@angular-eslint/builder": "15.1.0",
-    "@angular-eslint/eslint-plugin": "15.1.0",
-    "@angular-eslint/eslint-plugin-template": "15.1.0",
-    "@angular-eslint/schematics": "15.1.0",
-    "@angular-eslint/template-parser": "15.1.0",
-    "@angular/cli": "15.0.5",
-    "@angular/compiler-cli": "~15.0.4",
-    "@types/jasmine": "~4.3.0",
-    "@typescript-eslint/eslint-plugin": "5.44.0",
-    "@typescript-eslint/parser": "5.44.0",
-    "eslint": "^8.28.0",
-    "jasmine-core": "~4.5.0",
-    "karma": "~6.4.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.2.0",
-    "karma-jasmine": "~5.1.0",
-    "karma-jasmine-html-reporter": "~2.0.0",
-    "husky": "7.0.2",
-    "postcss": "^8.4.14",
-    "prettier": "2.4.1",
-    "prettier-eslint": "13.0.0",
-    "tailwindcss": "3.1.2",
-    "typescript": "4.8.4"
-  }
+    "name": "angular-boilerplate",
+    "version": "0.0.0",
+    "private": true,
+    "scripts": {
+        "postinstall": "npx simple-git-hooks",
+        "ng": "ng",
+        "start": "ng serve",
+        "build": "ng build",
+        "watch": "ng build --watch --configuration development",
+        "test": "ng test",
+        "test:headless": "ng test --watch=false --browsers ChromeHeadless",
+        "lint": "ng lint",
+        "lint:fix": "ng lint --fix",
+        "stylelint": "stylelint **/*.{css,pcss,less,scss,sass} --quiet-deprecation-warnings",
+        "stylelint:fix": "stylelint **/*.{css,pcss,less,scss,sass} --quiet-deprecation-warnings --fix"
+    },
+    "simple-git-hooks": {
+        "pre-commit": "pnpm lint-staged"
+    },
+    "lint-staged": {
+        "*.{js,jsx,ts,tsx,vue,svelte}": [
+            "pnpm lint:fix",
+            "pnpm prettier"
+        ],
+        "*.{svelte,vue,tsx,css,pcss,less,scss,sass}": [
+            "pnpm stylelint:fix",
+            "pnpm prettier"
+        ]
+    },
+    "dependencies": {
+        "@angular/animations": "^16.0.0",
+        "@angular/common": "^16.0.0",
+        "@angular/compiler": "^16.0.0",
+        "@angular/core": "^16.0.0",
+        "@angular/forms": "^16.0.0",
+        "@angular/platform-browser": "^16.0.0",
+        "@angular/platform-browser-dynamic": "^16.0.0",
+        "@angular/router": "^16.0.0",
+        "rxjs": "~7.8.0",
+        "tslib": "^2.3.0",
+        "zone.js": "~0.13.0"
+    },
+    "devDependencies": {
+        "@angular-devkit/build-angular": "^16.0.1",
+        "@angular-eslint/builder": "16.0.1",
+        "@angular-eslint/eslint-plugin": "16.0.1",
+        "@angular-eslint/eslint-plugin-template": "16.0.1",
+        "@angular-eslint/schematics": "16.0.1",
+        "@angular-eslint/template-parser": "16.0.1",
+        "@angular/cli": "~16.0.1",
+        "@angular/compiler-cli": "^16.0.0",
+        "@culur/stylelint-config": "^1.2.0",
+        "@types/jasmine": "~4.3.0",
+        "@typescript-eslint/eslint-plugin": "5.59.2",
+        "@typescript-eslint/parser": "5.59.2",
+        "autoprefixer": "^10.4.14",
+        "eslint": "^8.39.0",
+        "jasmine-core": "~4.6.0",
+        "karma": "~6.4.0",
+        "karma-chrome-launcher": "~3.2.0",
+        "karma-coverage": "~2.2.0",
+        "karma-jasmine": "~5.1.0",
+        "karma-jasmine-html-reporter": "~2.0.0",
+        "lint-staged": "^13.2.2",
+        "postcss": "^8.4.23",
+        "simple-git-hooks": "^2.8.1",
+        "stylelint": "^15.6.1",
+        "tailwindcss": "^3.3.2",
+        "typescript": "~5.0.2"
+    }
 }
