@@ -1,16 +1,17 @@
-module.exports = {
-  content: ['./src/**/*.{html,ts}', './projects/**/*.{html,ts}'],
-  darkMode: 'class',
-  theme: {
-    fontFamily: {
-      display: ['Oswald', 'sans-serif'],
-      body: ['Poppins', 'sans-serif'],
+/** @type {import('tailwindcss').Config} */
+export default {
+    content: ["./src/**/*.{html,ts}", "./projects/**/*.{html,ts}"],
+    darkMode: "class",
+    theme: {
+        fontFamily: {
+            display: ["Oswald", "sans-serif"],
+            body: ["Poppins", "sans-serif"],
+        },
+        container: {
+            center: true,
+            padding: "1.5rem",
+        },
+        extend: {},
     },
-    container: {
-      center: true,
-      padding: '1.5rem',
-    },
-    extend: {},
-  },
-  plugins: [],
+    plugins: [],
 };
