@@ -1,58 +1,62 @@
 import { inject } from '@angular/core';
-import { CanLoadFn, Route, Router, UrlSegment } from '@angular/router';
+import { CanMatchFn, Route, Router, UrlSegment } from '@angular/router';
 import { AuthService } from '@lib/services';
 
+type AuthGuardOptions = {
+    requiresAuthentication: boolean;
+};
+
+const defaultAuthGuardOptions = (): AuthGuardOptions => ({
+    requiresAuthentication: true,
+});
+
 /**
- * Creates a function that acts as an Angular route guard, allowing or blocking navigation based on the user's authentication status.
+ * Guard that allows or blocks the navigation based on the user's authentication status.
  *
- * @param options - An optional object that configures the behavior of the guard.
- * @param options.requiresAuthentication - An optional boolean that specifies whether the guard should allow or block navigation based on the user's authentication status. Defaults to `true`.
+ * @param options An optional object that configures the behavior of the guard.
+ * @param options.requiresAuthentication An optional boolean that specifies whether the guard should allow or block navigation based on the user's authentication status.
  *
  * @returns A function that acts as an Angular route guard.
  *
  * @example
- *
  * import { authGuard } from '@lib/guards';
  *
  * const routes: Routes = [
  *   {
  *     path: 'my-protected-component-path',
  *     loadComponent: async () => (await import('./path/to/my-protected-component')).MyProtectedComponent,
- *     canLoad: [authGuard()],
+ *     canMatch: [authGuard()],
  *   },
  * ];
  *
  * @example
- *
  * import { authGuard } from '@lib/guards';
  *
  * const routes: Routes = [
  *   {
  *     path: 'my-path-component',
  *     loadComponent: async () => (await import('./path/to/my-auth-component')).MyAuthComponent,
- *     canLoad: [authGuard({ requiresAuthentication: false })],
+ *     canMatch: [authGuard({ requiresAuthentication: false })],
  *   },
  * ];
  */
 export const authGuard = (
-  options: {
-    requiresAuthentication?: boolean;
-  } = { requiresAuthentication: true },
-): CanLoadFn => {
-  return (_: Route, segments: UrlSegment[]) => {
-    const router = inject(Router);
-    const authService = inject(AuthService);
+    options: AuthGuardOptions = defaultAuthGuardOptions()
+): CanMatchFn => {
+    return (_: Route, segments: UrlSegment[]) => {
+        const router = inject(Router);
+        const authService = inject(AuthService);
 
-    const isAuthenticated = authService.isAuthenticated;
+        if (options.requiresAuthentication === authService.isAuthenticated) {
+            return true;
+        }
 
-    if (options.requiresAuthentication === isAuthenticated) {
-      return true;
-    }
-
-    return options.requiresAuthentication
-      ? router.createUrlTree(['/auth/login'], {
-          queryParams: { returnUrl: segments.map((s) => s.path).join('/') },
-        })
-      : router.createUrlTree(['/']);
-  };
+        return options.requiresAuthentication
+            ? router.createUrlTree(['/auth/login'], {
+                  queryParams: {
+                      returnUrl: segments.map((s) => s.path).join('/'),
+                  },
+              })
+            : router.createUrlTree(['/']);
+    };
 };
